requirements.txt
amqp==5.0.1
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
numpy==1.19.2
pandas==1.1.3
prompt-toolkit==3.0.8
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
